pyproject.toml
[project]
name = "deepjscc"
version = "0.1.0"
description = "Deep Joint Source-Channel Coding (DeepJSCC) MVP"
readme = "README.md"
authors = [{ name = "dudosya", email = "kenaykay@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
]

[project.scripts]
deepjscc = "deepjscc.cli:app"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
